third_party/WebKit/LayoutTests/http/tests/inspector/console-xhr-logging-expected.txt

CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE WARNING: Setting 'XMLHttpRequest.withCredentials' for synchronous requests is deprecated.
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Tests that XMLHttpRequest Logging works when Enabled and doesn't show logs when Disabled.

console-xhr-logging.html:10 sending a GET request to resources/xhr-exists.html
network-test.js:129 Setting 'XMLHttpRequest.withCredentials' for synchronous requests is deprecated.
network-test.js:132 XHR finished loading: GET "http://127.0.0.1:8000/inspector/resources/xhr-exists.html".
console-xhr-logging.html:10 sending a GET request to resources/xhr-does-not-exist.html
network-test.js:132 GET http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html 404 (Not Found)
network-test.js:132 XHR finished loading: GET "http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html".
console-xhr-logging.html:10 sending a POST request to resources/post-target.cgi
network-test.js:132 XHR finished loading: POST "http://127.0.0.1:8000/inspector/resources/post-target.cgi".
console-xhr-logging.html:10 sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html
network-test.js:132 XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
console-xhr-logging.html:10 sending a GET request to resources/xhr-exists.html
console-xhr-logging.html:10 sending a GET request to resources/xhr-does-not-exist.html
network-test.js:132 GET http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html 404 (Not Found)
console-xhr-logging.html:10 sending a POST request to resources/post-target.cgi
console-xhr-logging.html:10 sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html
network-test.js:132 XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
